requirements.txt
streamlit>=1.30.0
tensorflow==2.18.0
opencv-python-headless>=4.5.5
Pillow>=9.5.0
numpy>=1.24.0
